requirements.txt
numpy==1.26.4
tensorflow-datasets==4.9.3
tensorflow==2.16.1
PyYAML==6.0.1
scikit-learn==1.4.2
matplotlib==3.9.0
pandas==2.2.1
pytest==8.2.0
